+++ tests/test_merkle_block.cpp
@@ -316,7 +316,7 @@ static bool B01_header_size() {
     hdr.version = 1;
     hdr.prev_block_hash = MakeHash(0xAA);
     hdr.merkle_root = MakeHash(0xBB);
-    hdr.timestamp = 1773619200;
+    hdr.timestamp = 1773597600;
     hdr.bits_q = GENESIS_BITSQ;
     hdr.nonce = 42;
     hdr.height = 0;
@@ -336,7 +336,7 @@ static bool B02_header_roundtrip() {
     hdr.version = 1;
     hdr.prev_block_hash = MakeHash(0x11);
     hdr.merkle_root = MakeHash(0x22);
-    hdr.timestamp = 1773619200;
+    hdr.timestamp = 1773597600;
     hdr.bits_q = GENESIS_BITSQ;
     hdr.nonce = 999999;
     hdr.height = 1000;
@@ -359,7 +359,7 @@ static bool B03_block_hash_deterministic() {
     hdr.version = 1;
     hdr.prev_block_hash = MakeHash(0xAA);
     hdr.merkle_root = MakeHash(0xBB);
-    hdr.timestamp = 1773619200;
+    hdr.timestamp = 1773597600;
     hdr.bits_q = GENESIS_BITSQ;
     hdr.nonce = 42;
     hdr.height = 0;
@@ -385,7 +385,7 @@ static bool B03_block_hash_deterministic() {
 static bool B04_hash_manual_verification() {
     BlockHeader hdr;
     hdr.version = 1;
-    hdr.timestamp = 1773619200;
+    hdr.timestamp = 1773597600;
     hdr.bits_q = GENESIS_BITSQ;
 
     auto serialized = hdr.Serialize();
@@ -409,7 +409,7 @@ static bool B05_deserialize_short() {
 // B06: Genesis header helper
 static bool B06_genesis_header() {
     Hash256 root = MakeHash(0xAA);
-    int64_t ts = 1773619200;
+    int64_t ts = 1773597600;
     uint32_t bits = GENESIS_BITSQ;
 
     BlockHeader hdr = MakeGenesisHeader(root, ts, bits);
@@ -480,7 +480,7 @@ static bool B08_le_encoding() {
 static bool B09_hash_hex() {
     BlockHeader hdr;
     hdr.version = 1;
-    hdr.timestamp = 1773619200;
+    hdr.timestamp = 1773597600;
     hdr.bits_q = GENESIS_BITSQ;
 
     std::string hex = hdr.ComputeBlockHashHex();
@@ -528,7 +528,7 @@ static bool B12_block_roundtrip() {
 
     Block blk;
     blk.header.version = 1;
-    blk.header.timestamp = 1773619200;
+    blk.header.timestamp = 1773597600;
     blk.header.bits_q = GENESIS_BITSQ;
     blk.header.height = 0;
 
@@ -681,7 +681,7 @@ static bool B18_hash_chain() {
     blk0.version = 1;
     blk0.prev_block_hash = Hash256{};
     blk0.merkle_root = MakeHash(0x11);
-    blk0.timestamp = 1773619200;
+    blk0.timestamp = 1773597600;
     blk0.bits_q = GENESIS_BITSQ;
     blk0.nonce = 42;
     blk0.height = 0;
@@ -720,7 +720,7 @@ static bool I01_full_block_lifecycle() {
     // Build block
     Block blk;
     blk.header.version = 1;
-    blk.header.timestamp = 1773619200;
+    blk.header.timestamp = 1773597600;
     blk.header.bits_q = GENESIS_BITSQ;
     blk.header.height = 0;
 
@@ -759,7 +759,7 @@ static bool I02_two_block_chain() {
     Block blk0;
     blk0.header.version = 1;
     blk0.header.prev_block_hash = Hash256{};
-    blk0.header.timestamp = 1773619200;
+    blk0.header.timestamp = 1773597600;
     blk0.header.bits_q = GENESIS_BITSQ;
     blk0.header.height = 0;
     blk0.txs.push_back(MakeTestCoinbase(0, 785100863, 0,
@@ -803,7 +803,7 @@ static bool I03_serialize_deserialize_verify() {
 
     Block blk;
     blk.header.version = 1;
-    blk.header.timestamp = 1773619200;
+    blk.header.timestamp = 1773597600;
     blk.header.bits_q = GENESIS_BITSQ;
     blk.header.nonce = 12345;
     blk.header.height = 42;